requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.1
cryptography==2.3.0
cycler==0.10.0
decorator==4.4.0
enum34==1.1.6
future==0.17.1
idna==2.8
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
matplotlib==2.2.4
networkx==2.2
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.1.0
pycrypto==2.6.1
pygobject==3.32.2
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pyxdg==0.26
PyYAML==5.1
requests==2.22.0
scikit-image==0.14.3
scikit-video==1.1.11
scipy==1.2.2
SecretStorage==2.3.1
six==1.12.0
subprocess32==3.5.4
torch==1.1.0
tqdm==4.32.2
urllib3==1.25.3
